test: ensure ChangeProfilePicture not call DeleteFile if file does not exists and SaveUserPicture throws

+++ src/domain/use-cases/change-profile-picture.ts
@@ -34,7 +34,8 @@ export const setupChangeProfilePicture: Setup =
     try {
       await userProfileRepo.savePicture(userProfile);
     } catch {
-      await fileStorage.delete({ key });
+      if (file !== undefined) await fileStorage.delete({ key });
+      throw new Error();
     }
     return userProfile;
   };
